requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
asyncer==0.0.2
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.2
bidict==0.22.1
bs4==0.0.1
build==1.0.3
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.0
curl-cffi==0.5.10
cycler==0.12.1
dataclasses-json==0.5.14
deprecated==1.2.14
duckduckgo-search==4.1.1
fastapi==0.100.1
fastapi-socketio==0.0.10
ffmpy==0.3.1
filelock==3.13.1
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.47.0
frozenlist==1.4.1
fsspec==2023.12.2
google-auth==2.26.1
googleapis-common-protos==1.62.0
gradio==4.13.0
gradio-client==0.8.0
greenlet==3.0.3
grpcio==1.60.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.1
httpx==0.24.1
huggingface-hub==0.20.2
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kubernetes==28.1.0
langchain==0.0.354
langchain-community==0.0.8
langchain-core==0.1.6
langsmith==0.0.77
lazify==0.4.0
lxml==5.0.0
markdown-it-py==3.0.0
markupsafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
mdurl==0.1.2
mmh3==4.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
numpy==1.26.3
oauthlib==3.2.2
onnxruntime==1.16.3
openai==1.39.0
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
orjson==3.9.10
overrides==7.4.0
packaging==23.2
pandas==2.1.4
pillow==10.2.0
pip==23.3.1
posthog==3.1.0
protobuf==4.25.1
pulsar-client==3.4.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.1
pydantic-core==2.14.3
pydub==0.25.1
pygments==2.17.2
pyjwt==2.8.0
pyparsing==3.1.1
pypdf==3.17.4
pypika==0.48.9
pyproject-hooks==1.0.0
pyprojroot==0.3.0
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.1
python-graphql-client==0.4.3
python-multipart==0.0.6
python-socketio==5.10.0
pytz==2023.3.post1
pyyaml==6.0.1
referencing==0.32.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.16.2
rsa==4.9
semantic-version==2.10.0
setuptools==68.2.2
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlalchemy==2.0.25
starlette==0.27.0
sympy==1.12
syncer==2.0.3
tenacity==8.2.3
tiktoken==0.5.2
tokenizers==0.15.0
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.0
tqdm==4.66.1
typer==0.9.0
typing-extensions==4.9.0
typing-inspect==0.9.0
tzdata==2023.4
uptrace==1.22.0
urllib3==1.26.18
uvicorn==0.23.2
watchfiles==0.20.0
websocket-client==1.7.0
websockets==11.0.3
wheel==0.41.2
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0